Refactor OpenAI backend components to improve response handling and traceability. Introduce FinalResponse model for structured output, enhance HooksState management, and update OpenAITeamAdapter and OpenAIBackend for better integration with new features. Adjust SupervisorHooks to utilize final response formatting.
