+++ inline_agents/backends/openai/entities.py
@@ -1,6 +1,48 @@
 from dataclasses import dataclass
+from typing import Any, Dict, List
+
 from agents import Session
-from inline_agents.backends.openai.hooks import HooksState
+from pydantic import BaseModel, Field
+
+
+class HooksState:
+    def __init__(self, agents: list):
+        self.agents = agents
+        self.agents_names = []
+        self.lambda_names = {}
+        self.tool_calls = {}
+        self.trace_data = []
+        self.tool_info = {}
+
+        for agent in self.agents:
+            self.agents_names.append(agent.get("agentName"))
+            for action_group in agent.get("actionGroups", []):
+                action_group_name = action_group.get("actionGroupName")
+                function_names = []
+                for function_schema in action_group.get("functionSchema", {}).get("functions", []):
+                    function_name = function_schema.get("name")
+                    function_names.append(function_name)
+                self.lambda_names[action_group_name] = {
+                    "function_name": function_names[0],
+                    "function_arn": action_group.get("actionGroupExecutor", {}).get("lambda")
+                }
+
+    def add_tool_info(self, tool_name: str, info: Dict[str, Any]):
+        try:
+            self.tool_info[tool_name].update(info)
+        except KeyError:
+            self.tool_info[tool_name] = info
+
+    def add_tool_call(self, tool_call: Dict[str, Any]):
+        self.tool_calls.update(tool_call)
+
+    def get_events(self, result: dict, tool_name: str):
+        session_events = self.tool_info.get(tool_name, {}).get("events", {})
+        if session_events:
+            return session_events
+
+        events = result.get("events", {})
+        return events
 
 
 @dataclass
@@ -13,3 +55,8 @@ class Context:
     content_base: dict
     session: Session
     hooks_state: HooksState
+
+
+class FinalResponse(BaseModel):
+    """Modelo para a resposta final formatada"""
+    final_response: str = Field(description="O resultado final da resposta que ira ser formatado")
